Bugfix and change to show on hover.
The script used to add the rating for my avatar, next to comment box for example. Also it was a little hard to hover over the rating bar, though it was necessary in order to see the details.
Also I think it's less-intrusive and more usable when rating is shown only when hovering the thumbnail, or the video item.